requirements-x86_64.txt
Package                 Version
----------------------- ------------------
absl-py                 1.4.0
addict                  2.4.0
aiohttp                 3.8.4
aiosignal               1.3.1
albumentations          1.3.0
antlr4-python3-runtime  4.8
appdirs                 1.4.4
asttokens               2.2.1
async-timeout           4.0.2
attrs                   22.2.0
backcall                0.2.0
black                   23.1.0
cachetools              5.3.0
certifi                 2022.12.7
charset-normalizer      3.0.1
click                   8.1.3
cloudpickle             2.2.1
comm                    0.1.2
ConfigArgParse          1.5.3
contourpy               1.0.7
cycler                  0.11.0
dash                    2.8.1
dash-core-components    2.0.0
dash-html-components    2.0.0
dash-table              5.0.0
debugpy                 1.6.6
decorator               5.1.1
detectron2              0.6
distro                  1.8.0
docker-pycreds          0.4.0
executing               1.2.0
fastjsonschema          2.16.2
fire                    0.5.0
Flask                   2.2.3
fonttools               4.38.0
frozenlist              1.3.3
fsspec                  2023.1.0
fvcore                  0.1.5.post20221221
gitdb                   4.0.10
GitPython               3.1.31
google-auth             2.16.0
google-auth-oauthlib    0.4.6
grpcio                  1.51.1
hydra-core              1.0.5
idna                    3.4
imageio                 2.25.1
importlib-metadata      3.10.1
iniconfig               2.0.0
iopath                  0.1.10
ipykernel               6.21.2
ipython                 8.10.0
ipywidgets              8.0.4
itsdangerous            2.1.2
jedi                    0.18.2
Jinja2                  3.1.2
joblib                  1.2.0
jsonschema              4.17.3
jupyter_client          8.0.3
jupyter_core            5.2.0
jupyterlab-widgets      3.0.5
kiwisolver              1.4.4
lightning-utilities     0.6.0.post0
loguru                  0.6.0
Markdown                3.4.1
MarkupSafe              2.1.2
matplotlib              3.7.0
matplotlib-inline       0.1.6
MinkowskiEngine         0.5.4
multidict               6.0.4
mypy                    0.991
mypy-extensions         1.0.0
natsort                 8.2.0
nbformat                5.5.0
nest-asyncio            1.5.6
networkx                3.0
ninja                   1.10.2.3
numpy                   1.23.2
oauthlib                3.2.2
omegaconf               2.0.6
open3d                  0.16.0
opencv-python-headless  4.7.0.68
packaging               23.0
pandas                  1.4.2
parso                   0.8.3
pathspec                0.11.0
pathtools               0.1.2
pexpect                 4.8.0
pickleshare             0.7.5
Pillow                  9.4.0
pip                     22.3.1
platformdirs            3.0.0
plotly                  5.13.0
pluggy                  1.0.0
plyfile                 0.7.4
pointnet2               0.0.0
portalocker             2.7.0
prompt-toolkit          3.0.37
protobuf                4.22.0
psutil                  5.9.4
ptyprocess              0.7.0
pure-eval               0.2.2
py                      1.11.0
pyasn1                  0.4.8
pyasn1-modules          0.2.8
pycocotools             2.0.6
pyDeprecate             0.3.2
Pygments                2.14.0
PyOpenGL                3.1.6
PyOpenGL-accelerate     3.1.6
pyparsing               3.0.9
PyQt5                   5.15.9
PyQt5-Qt5               5.15.2
PyQt5-sip               12.11.1
PyQt5-stubs             5.15.6.0
pyquaternion            0.9.9
pyrsistent              0.19.3
pytest                  7.1.3
pytest-qt               4.1.0
python-dateutil         2.8.2
python-dotenv           0.21.1
pytorch-lightning       1.7.2
pytz                    2022.7.1
pyviz3d                 0.2.32
PyWavelets              1.4.1
PyYAML                  5.4.1
pyzmq                   25.0.0
qudida                  0.0.4
regex                   2022.10.31
requests                2.28.2
requests-oauthlib       1.3.1
rsa                     4.9
scikit-image            0.19.3
scikit-learn            1.2.1
scipy                   1.10.0
sentry-sdk              1.15.0
setproctitle            1.3.2
setuptools              65.6.3
six                     1.16.0
smbus2                  0.4.2
smmap                   5.0.0
stack-data              0.6.2
tabulate                0.9.0
tenacity                8.2.1
tensorboard             2.12.0
tensorboard-data-server 0.7.0
tensorboard-plugin-wit  1.8.1
termcolor               2.2.0
threadpoolctl           3.1.0
tifffile                2023.2.3
toml                    0.10.2
tomli                   2.0.1
torch                   1.12.1+cu116
torch-scatter           2.0.9
torchmetrics            0.11.1
torchvision             0.13.1+cu116
tornado                 6.2
tqdm                    4.64.1
traitlets               5.9.0
trimesh                 3.20.0
types-pkg-resources     0.1.3
types-setuptools        57.4.18
typing_extensions       4.5.0
urllib3                 1.26.14
volumentations          0.1.8
wandb                   0.13.10
wcwidth                 0.2.6
Werkzeug                2.2.3
wheel                   0.38.4
widgetsnbextension      4.0.5
yacs                    0.1.8
yarl                    1.8.2
zipp                    3.13.0